use -Wall in walking

+++ bitbots_quintic_walk/src/walk_ik.cpp
@@ -66,7 +66,7 @@ void WalkIK::reset() {
   // based method. Otherwise, the first step will be not correct
   std::vector<std::string> names_vec = {"LHipPitch", "LKnee", "LAnklePitch", "RHipPitch", "RKnee", "RAnklePitch"};
   std::vector<double> pos_vec = {0.7, 1.0, -0.4, -0.7, -1.0, 0.4};
-  for (long i = 0; i < names_vec.size(); i++) {
+  for (size_t i = 0; i < names_vec.size(); i++) {
     // besides its name, this method only changes a single joint position...
     goal_state_->setJointPositions(names_vec[i], &pos_vec[i]);
   }
